kodekaleesh-2025-main/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.0.0
boto3==1.26.137
pytest==7.4.3
pytest-cov==4.1.0
